feat: wishlist order update 기능 추가

+++ onboarding_app/queries/wishlist.py
@@ -19,7 +19,7 @@ def create_wishlist(
             user_id=current_user.id,
             name=wishlist.name,
             description=wishlist.description,
-            order_num=count_for_order + 1,
+            order_num=count_for_order,
         )
         db.add(db_wishlist)
         db.commit()
@@ -58,7 +58,7 @@ def fetch_wishlists(
     elif order_by == "ordered":
         # TODO: add order-by order_num
         return (
-            db_wishlists.order_by(models.Wishlist.order_num.desc())
+            db_wishlists.order_by(models.Wishlist.order_num)
             .limit(limit)
             .offset(offset)
             .all()
@@ -78,6 +78,17 @@ def get_wishlist(
     elif db_wishlist.first().user_id != current_user.id:
         print(db_wishlist.first().user_id, current_user.id)
         raise exceptions.PermissionDeniedError
+    return db_wishlist.first()
+
+
+def get_wishlist_by_name(
+    db: Session, name: str, current_user: schemas.User
+) -> models.Wishlist:
+    db_wishlist = db.query(models.Wishlist).filter(
+        models.Wishlist.user_id == current_user.id, models.Wishlist.name == name
+    )
+    if not db_wishlist.first():
+        raise exceptions.DataDoesNotExistError
     return db_wishlist.first()
 
 
@@ -111,5 +122,54 @@ def delete_wishlist(
     elif db_wishlist.first().user_id != current_user.id:
         raise exceptions.PermissionDeniedError
     db_wishlist.delete()
+    users_db_wishlist = (
+        db.query(models.Wishlist)
+        .filter(models.Wishlist.user_id == current_user.id)
+        .order_by(models.Wishlist.order_num)
+    )
+    for i, wishlist in enumerate(users_db_wishlist):
+        wishlist.order_num = i
+    db.commit()
+    return db_wishlist.first()
+
+
+def change_wishlist_order(
+    db: Session,
+    current_user: schemas.User,
+    wishlist_id: int,
+    hope_order: int,
+) -> models.Wishlist:
+    db_wishlist = db.query(models.Wishlist).filter(models.Wishlist.id == wishlist_id)
+    users_db_wishlist = db.query(models.Wishlist).filter(
+        models.Wishlist.user_id == current_user.id
+    )
+
+    if not db_wishlist.first():
+        raise exceptions.DataDoesNotExistError
+    elif db_wishlist.first().user_id != current_user.id:
+        raise exceptions.PermissionDeniedError
+    elif hope_order < 0 or hope_order >= users_db_wishlist.count():
+        raise exceptions.InvalidQueryError
+
+    if hope_order > db_wishlist.first().order_num:
+        db.query(models.Wishlist).filter(
+            models.Wishlist.order_num > db_wishlist.first().order_num,
+            models.Wishlist.order_num <= hope_order,
+            models.Wishlist.user_id == current_user.id,
+        ).update(
+            {models.Wishlist.order_num: models.Wishlist.order_num - 1},
+            synchronize_session=False,
+        )
+
+    elif hope_order < db_wishlist.first().order_num:
+        db.query(models.Wishlist).filter(
+            models.Wishlist.order_num < db_wishlist.first().order_num,
+            models.Wishlist.order_num >= hope_order,
+            models.Wishlist.user_id == current_user.id,
+        ).update(
+            {models.Wishlist.order_num: models.Wishlist.order_num + 1},
+            synchronize_session=False,
+        )
+    db_wishlist.first().order_num = hope_order
     db.commit()
     return db_wishlist.first()
